+++ CHANGELOG.md
@@ -8,7 +8,7 @@ v0.10.4 - TBD
 - **Fixed:** [#55](https://github.com/Abscissa/SDLang-D/issues/55):
 	TimeZone.getTimeZone deprecation hotfix (@ZILtoid1991)
 - **Tests:** Updated compiler list.
-- **Tests:** Fix tests on DMD 2.081 by updating unit-threaded.
+- **Tests:** Fix tests on DMD 2.081 by using an updated unit-threaded.
 - **Tests:** Test on OSX.
 
 v0.10.3 - 2018-05-30